@@ -38,8 +38,8 @@
     "ejs": "^3.1.9",
     "etherpad-require-kernel": "^1.0.16",
     "etherpad-yajsml": "0.0.12",
-    "express": "4.18.2",
-    "express-rate-limit": "^7.1.5",
+    "express": "4.18.3",
+    "express-rate-limit": "^7.2.0",
     "express-session": "npm:@etherpad/express-session@^1.18.2",
     "fast-deep-equal": "^3.1.3",
     "find-root": "1.1.0",
@@ -63,13 +63,13 @@
     "security": "1.0.0",
     "semver": "^7.6.0",
     "socket.io": "^3.1.2",
-    "socket.io-client": "^4.7.4",
+    "socket.io-client": "^3.1.2",
     "superagent": "^8.1.2",
-    "terser": "^5.28.1",
+    "terser": "^5.29.1",
     "threads": "^1.7.0",
     "tinycon": "0.6.8",
     "tsx": "^4.7.1",
-    "ueberdb2": "^4.2.57",
+    "ueberdb2": "^4.2.60",
     "underscore": "1.13.6",
     "unorm": "1.6.0",
     "wtfnode": "^0.9.1"
@@ -84,13 +84,13 @@
     "@types/http-errors": "^2.0.4",
     "@types/jsdom": "^21.1.6",
     "@types/mocha": "^10.0.6",
-    "@types/node": "^20.11.20",
+    "@types/node": "^20.11.25",
     "@types/sinon": "^17.0.3",
     "@types/supertest": "^6.0.2",
     "@types/semver": "^7.5.7",
     "@types/underscore": "^1.11.15",
     "cypress": "^13.6.4",
-    "eslint": "^8.56.0",
+    "eslint": "^8.57.0",
     "eslint-config-etherpad": "^3.0.22",
     "etherpad-cli-client": "^3.0.1",
     "mocha": "^10.3.0",
@@ -102,7 +102,7 @@
     "sinon": "^17.0.1",
     "split-grid": "^1.0.11",
     "supertest": "^6.3.4",
-    "typescript": "^5.3.3"
+    "typescript": "^5.4.2"
   },
   "engines": {
     "node": ">=18.18.2",
