@@ -2,12 +2,12 @@
 
 import {ArgsExpressType} from "../../types/ArgsExpressType";
 
-const events = require('events');
+import events from 'events';
 const express = require('../express');
-const log4js = require('log4js');
+import log4js from 'log4js';
 const proxyaddr = require('proxy-addr');
 const settings = require('../../utils/Settings');
-import {Server} from 'socket.io'
+import {Server, Socket} from 'socket.io'
 const socketIORouter = require('../../handler/SocketIORouter');
 const hooks = require('../../../static/js/pluginfw/hooks');
 const padMessageHandler = require('../../handler/PadMessageHandler');
@@ -17,7 +17,7 @@ const logger = log4js.getLogger('socket.io');
 const sockets = new Set();
 const socketsEvents = new events.EventEmitter();
 
-exports.expressCloseServer = async () => {
+export const expressCloseServer = async () => {
   if (io == null) return;
   logger.info('Closing socket.io engine...');
   // Close the socket.io engine to disconnect existing clients and reject new clients. Don't call
@@ -48,7 +48,7 @@ exports.expressCloseServer = async () => {
   logger.info('All socket.io clients have disconnected');
 };
 
-exports.socketSessionMiddleware = (args: any) => (socket: any, next: Function) => {
+const socketSessionMiddleware = (args: any) => (socket: any, next: Function) => {
   const req = socket.request;
   // Express sets req.ip but socket.io does not. Replicate Express's behavior here.
   if (req.ip == null) {
@@ -65,59 +65,35 @@ exports.socketSessionMiddleware = (args: any) => (socket: any, next: Function) =
   express.sessionMiddleware(req, {}, next);
 };
 
-exports.expressCreateServer = (hookName:string, args:ArgsExpressType, cb:Function) => {
+export const expressCreateServer = (hookName:string, args:ArgsExpressType, cb:Function) => {
   // init socket.io and redirect all requests to the MessageHandler
   // there shouldn't be a browser that isn't compatible to all
   // transports in this list at once
   // e.g. XHR is disabled in IE by default, so in IE it should use jsonp-polling
-  io = new Server(args.server, {
+  io = new Server(args.server,{
     transports: settings.socketTransportProtocols,
-  }).listen(args.server, {
-    /*
-     * Do not set the "io" cookie.
-     *
-     * The "io" cookie is created by socket.io, and its purpose is to offer an
-     * handle to perform load balancing with session stickiness when the library
-     * falls back to long polling or below.
-     *
-     * In Etherpad's case, if an operator needs to load balance, he can use the
-     * "express_sid" cookie, and thus "io" is of no use.
-     *
-     * Moreover, socket.io API does not offer a way of setting the "secure" flag
-     * on it, and thus is a liability.
-     *
-     * Let's simply nuke "io".
-     *
-     * references:
-     *   https://socket.io/docs/using-multiple-nodes/#Sticky-load-balancing
-     *   https://github.com/socketio/socket.io/issues/2276#issuecomment-147184662 (not totally true, actually, see above)
-     */
     cookie: false,
     maxHttpBufferSize: settings.socketIo.maxHttpBufferSize,
+  })
+
+  io.on('connection', (socket:any) => {
+    sockets.add(socket);
+    socketsEvents.emit('updated');
+    // https://socket.io/docs/v3/faq/index.html
+    const session = socket.request.session;
+    session.connections++;
+    session.save();
+    socket.on('disconnect', () => {
+      sockets.delete(socket);
+      socketsEvents.emit('updated');
+    });
   });
 
-  function handleConnection() {
-    return (socket: any) => {
-      sockets.add(socket);
-      socketsEvents.emit('updated');
-      // https://socket.io/docs/v3/faq/index.html
-      const session = socket.request.session;
-      session.connections++;
-      session.save();
-      socket.on('disconnect', () => {
-        sockets.delete(socket);
-        socketsEvents.emit('updated');
-      });
-    };
-  }
-
-  io.on('connection', handleConnection);
-
-  io.use(exports.socketSessionMiddleware(args));
+  io.use(socketSessionMiddleware(args));
 
   // Temporary workaround so all clients go through middleware and handle connection
-  io.of('/pluginfw/installer').use(exports.socketSessionMiddleware(args))
-  io.of('/settings').use(exports.socketSessionMiddleware(args))
+  io.of('/pluginfw/installer').use(socketSessionMiddleware(args))
+  io.of('/settings').use(socketSessionMiddleware(args))
 
   io.use((socket:any, next:Function) => {
     socket.conn.on('packet', (packet:string) => {
